+++ firmware/controllers/sensors/allsensors.cpp
@@ -9,11 +9,13 @@
 
 #include "pch.h"
 
-ButtonDebounce acDebounce("ac_switch");
+static ButtonDebounce acDebounce("ac_switch");
+static ButtonDebounce acPressureDebounce("ac_pressure_switch");
 
 void initSensors() {
 	initMapDecoder();
 	acDebounce.init(MS2NT(15), engineConfiguration->acSwitch, engineConfiguration->acSwitchMode);
+	acPressureDebounce.init(MS2NT(15), engineConfiguration->acPressureSwitch, engineConfiguration->acPressureSwitchMode);
 }
 
 bool getAcToggle() {
@@ -22,4 +24,12 @@ bool getAcToggle() {
 
 bool hasAcToggle() {
 	return isBrainPinValid(engineConfiguration->acSwitch);
+}
+
+bool hasAcPressure() {
+	return isBrainPinValid(engineConfiguration->acPressureSwitch);
+}
+
+bool getAcPressure() {
+	return acPressureDebounce.readPinState();
 }